Refactor star image databinding logic using only one function

+++ app/src/main/java/com/sju18001/petmanagement/ui/map/review/ReviewListAdapter.kt
@@ -81,7 +81,7 @@ class ReviewListAdapter(
 
         setAccountPhoto(holder, data.author)
         Util.setViewMore(holder.contentsText, holder.viewMoreText, 3)
-        Util.setRatingStars(getStarImages(holder), data.rating.toFloat(), context)
+        Util.setRatingStars(getStarImages(holder), data.rating.toDouble(), context)
     }
 
     private fun setAccountPhoto(holder: ViewHolder, author: Account) {
